@@ -1,12 +1,11 @@
 apply plugin: 'com.android.library'
+apply plugin: 'maven-publish'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
-
+    compileSdkVersion 34
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 23
+        targetSdkVersion 34
         versionCode 33
         versionName "2.1.3"
     }
@@ -19,18 +18,52 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
 }
 
-ext {
-    PUBLISH_GROUP_ID = 'com.wang.avi'
-    PUBLISH_VERSION = '2.1.3'
-    PUBLISH_DES = 'Nice loading animations for Android'
-    LIB_NAME = 'AVLoadingIndicatorView'
+afterEvaluate {
+    publishing {
+        publications {
+            mavenJava(MavenPublication) {
+                groupId = "com.wang.avi"
+                artifactId = "library"
+                version = "2.1.3"
 
-    WEBSITE_URL = 'https://github.com/81813780/AVLoadingIndicatorView'
-    ISSUE_TRACKER_URL = 'https://github.com/81813780/AVLoadingIndicatorView/issues'
-    VSC_URL = 'https://github.com/81813780/AVLoadingIndicatorView'
+                // Include the AAR file manually
+                artifact("$buildDir/outputs/aar/${project.name}-release.aar") {
+                    builtBy tasks.named("assembleRelease")
+                }
+
+                pom {
+                    name = 'AVLoadingIndicatorView'
+                    description = 'Nice loading animations for Android'
+                    url = 'https://github.com/valid8me-engineering/AVLoadingIndicatorView'
+                    licenses {
+                        license {
+                            name = 'The Apache License, Version 2.0'
+                            url = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
+                        }
+                    }
+                    developers {
+                        developer {
+                            id = 'Shay-valid8Me'
+                            name = 'Shay Brennan-Kelly'
+                            email = 'sbrennankelly@valid8me.com'
+                        }
+                    }
+                }
+            }
+        }
+
+        repositories {
+            maven {
+                name = "GitHubPackages"
+                url = "https://maven.pkg.github.com/valid8me-engineering/AVLoadingIndicatorView"
+                credentials {
+                    username = System.getenv("GITHUB_ACTOR") ?: "default_actor"
+                    password = System.getenv("GITHUB_TOKEN") ?: "default_token"
+                }
+            }
+        }
+    }
 }
-
-apply from: 'https://raw.githubusercontent.com/andforce/release-android-lib-to-jcenter/master/bintray.gradle'
